+++ pom.xml
@@ -9,7 +9,6 @@
   <version>1.0-SNAPSHOT</version>
 
   <name>finaLP</name>
-  <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
 
   <properties>
@@ -24,6 +23,13 @@
       <artifactId>junit</artifactId>
       <version>4.11</version>
       <scope>test</scope>
+    </dependency>
+
+    <!-- flatlaf dependecia -->
+    <dependency>
+      <groupId>com.formdev</groupId>
+      <artifactId>flatlaf</artifactId>
+      <version>2.6</version>
     </dependency>
   </dependencies>
