+++ src/main/java/src/bankServer/regsYcomprobs/DatosComprobante.java
@@ -15,15 +15,10 @@ public class DatosComprobante {
     public String metodo;
     public String tipos;
 
-    public DatosComprobante(
-            Integer id,
-            String concepto,
-            Integer monto,
-            LocalDate fecha,
-            Cuenta cuentaOrigen,
-            Cuenta cuentaDestino,
-            String metodo,
-            String tipos) {
+    public DatosComprobante(Integer id, String concepto, Integer monto,
+            LocalDate fecha, Cuenta cuentaOrigen, Cuenta cuentaDestino,
+            String metodo, String tipos) {
+
         this.id = id;
         this.concepto = concepto;
         this.cuentaDestino = cuentaDestino;
@@ -33,4 +28,6 @@ public class DatosComprobante {
         this.tipos = tipos;
         this.fecha = fecha;
     }
+
+    public DatosComprobante () { }
 }
